+++ src/composables/subscribe.ts
@@ -1,6 +1,7 @@
 import { RouteLocationRaw } from "vue-router"
 
 import { Profile } from "@/api/users"
+import { useInstanceInfo } from "@/composables/instance"
 
 interface SubscriptionOption {
   type: "ethereum" | "monero",
@@ -8,6 +9,8 @@ interface SubscriptionOption {
 }
 
 export function useSubscribe() {
+  const { getBlockchainInfo } = useInstanceInfo()
+
   function getSubscriptionOption(profile: Profile): SubscriptionOption | null {
     for (const option of profile.payment_options) {
       if (
@@ -23,6 +26,11 @@ export function useSubscribe() {
         option.type === "ethereum-subscription" ||
         option.type === "monero-subscription"
       ) {
+        const blockchain = getBlockchainInfo()
+        if (!blockchain?.features.subscriptions) {
+          // Subscription option disabled
+          continue
+        }
         return {
           type: option.type === "ethereum-subscription" ? "ethereum" : "monero",
           location: {