+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Show subscription button if blockchain integration is disabled.
+
 ## [1.23.0] - 2023-05-03
 
 ### Added